ci: add cppcheck and clang-tidy static analysis to CI
Adds two new GitHub Actions jobs running on every push and PR:
- cppcheck: warning/style/performance/portability with curated
  suppressions for false positives (Qt MOC, intentional bitmask init)
- clang-tidy: bugprone, cert, performance, portability with platform
  files excluded (STM32 HAL, Zephyr kernel, LwIP raw API don't build
  on Linux)

Source fixes applied for 6 real findings:
- duplicate expression in ternary (zephyr main.cpp)
- const-correctness on stack arrays (zephyr main.cpp)
- uninitialized pbuf pointers annotated (LwIP RX paths)
- implicit widening in SD codec
- discarded return value in operator service consumer
- branch-clone arms reviewed, fixed or annotated

Suppressions document deferred items inline; configuration follows
automotive-leaning policy (strict on bugprone-* and cert-*, relaxed on
style and naming). Production-grade projects layer MISRA-C++ tools
(Polyspace, Coverity, LDRA) on top of this baseline.

+++ body_control_zonal_lighting/src/transport/some_ip_sd/some_ip_sd_codec.cpp
@@ -241,7 +241,7 @@ bool DecodeOffer(
 
     for (std::uint32_t i = 0U; i < num_entries; ++i)
     {
-        const std::uint8_t* const entry = sd + 8U + (i * kEntrySize);
+        const std::uint8_t* const entry = sd + 8U + (static_cast<std::size_t>(i) * kEntrySize);
         if (entry[0] != kEntryTypeOffer) { continue; }  // not OfferService
 
         const std::uint16_t svc  = ReadU16Be(entry + 4U);
